+++ _define.php
@@ -2,7 +2,7 @@
 # -- BEGIN LICENSE BLOCK ----------------------------------
 # This file is part of rateIt, a plugin for Dotclear 2.
 #
-# Copyright(c) 2014 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
+# Copyright(c) 2014-2015 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
 #
 # Copyright (c) 2009-2010 JC Denis and contributors
 # jcdenis@gdwd.com
@@ -17,7 +17,14 @@ if (!defined('DC_RC_PATH')){return;}
 $this->registerModule(
 	/* Name */			"Rate it",
 	/* Description*/		"Rating system for your blogs",
-	/* Author */			"JC Denis",
-	/* Version */			'2.1.0',
-	/* Permissions */		'admin'
+	/* Author */			"JC Denis, Nicolas Roudaire",
+	/* Version */			'2.2.0',
+	/* Properties */
+	array(
+		'permissions' => 'admin',
+		'type' => 'plugin',
+		'dc_min' => '2.7',
+		'support' => 'http://forum.dotclear.org/viewtopic.php?id=39801',
+		'details' => 'http://plugins.dotaddict.org/dc2/details/rateIt'
+		)
 );